refactor: HistoryForm component
