--- src/components/HistoryForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from './HistoryForm';
